% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryConf.R
\name{boundaryConf}
\alias{boundaryConf}
\title{Calculate the upper bound or lower bound estimate of the 
confidence for the stratified sampling model}
\usage{
boundaryConf(N, n, B, ub, rho = 1, simpsonSample = 50000 + 1)
}
\arguments{
\item{N}{The sum of all grid cells within the population}

\item{n}{The sum of all sampled grid cells within the population}

\item{B}{The value of the hyperparameter associated with the beta prior}

\item{ub}{The number of unacceptable items one is willing to accept}
}
\value{
conf The confidence
}
\description{
Calculate the upper bound or lower bound estimate of the 
confidence for the stratified sampling model
}
\note{
To produce a lower bound estimate, set rho = min(rho).
To produce an upper bound estimate, set rho = 1.
}

