% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundedSS.R
\name{boundedSS}
\alias{boundedSS}
\title{Compute the upper bound or lower bound for stratified sample size}
\usage{
boundedSS(N, B, ub, targetConf, rho = 1)
}
\arguments{
\item{N}{The sum of all grid cells within the population}

\item{B}{The value of the hyperparameter associated with the beta prior}

\item{ub}{The number of unacceptable items one is willing to accept}

\item{targetConf}{The confidence that you want pct.clean of the grid cells
in the population to be acceptable}

\item{rho}{Determines whether you are computing the upper bound (rho = 1) or 
the lower bound (min(rho)) for sample size. By default, the upper bound is 
computed (rho = 1)}
}
\value{
\itemize{
\item{samples}{ The required sample size}
\item{achievedConf}{ The confidence achieved when taking samples}
\item{targetConf}{ The confidence you were trying to achieve}
}
}
\description{
Compute the upper bound or lower bound for stratified sample size
}

