% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confRaw.R
\name{confRaw}
\alias{confRaw}
\title{Compute the Bayesian confidence for the stratified sampling model}
\usage{
confRaw(n, rho, N, B, ub, subInts = 10001, method = c("simpson",
  "quadrature"), clusterName = NULL)
}
\arguments{
\item{n}{the vector of samples to take from each stratum}

\item{rho}{a vector of relative weights of finding an unacceptable item relative to
stratum 1. Note that rho[1] = 1 always.}

\item{N}{the vector of total grid cells for each stratum}

\item{B}{the beta hyperparameter of theta ~ beta(1, B)
larger values of B imply lower probability of finding an unacceptable item}

\item{ub}{the total number of unacceptable items one is willing to accept
that remain in the unsampled grid cells}

\item{subInts}{the number of sub-intervals to use for the application of
Simpson's rule. The number must be odd. Default = 10001.}

\item{method}{The integration method:  Simpson's rule or quadrature}

\item{clusterName}{The name of a cluster created by
\code{\link[parallel:makeCluster]{parallel::makeCluster}}, which,
if supplied will initiate parallel computation of the integrand before passing it into
Simpson's method.  Ignored for \code{method = "quadrature"}.  If \code{NULL}, no
parallelization is performed.}
}
\value{
the Bayesian confidence as estimated by an application of Simpson's Rule.
}
\description{
Compute the Bayesian confidence for the stratified sampling model
}
\examples{

# N must be an integer
N <- c(1000, 1000, 1000)

# n must be an integer
n <- c(100, 75, 50)

rho <- c(1, 0.85, 0.75)

B <- 70

# 1\% of total grid cells
ub <- floor(.01*sum(N))

confRaw(n, rho, N, B, ub)
}
\author{
Alex Venzin
}

