% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Find the minimum sample size that maximizes alpha}
\usage{
minmax(results, epsilon = 0.01)
}
\arguments{
\item{results}{the data.frame produced by \code{\link{findSS}}}

\item{epsilon}{the tolerance parameter for accepting a larger sample
size in exchange for a larger alpha. Choose larger sample size if 
\eqn{(min \sum n - candidate \sum n)/(min \sum n)
< \epsilon}}
}
\value{
The minimum sample allocation with the maximum alpha (more proportional sampling)
}
\description{
Find the minimum sample size that maximizes alpha
}
\examples{

# generate a candidate data frame of sample sizes 

sampleAllocations <- findSS(conf = 0.95, rho = c(1, 0.5), N = c(1000,2000),
                            prior.prob = 0.99, pct.clean = 0.99, 
                            alphaSeq = c(0.5, 0.8), minmax = FALSE, nJobs = 2)

# Return the allocation that minimizes sum n and maximizes alpha.
optimum <- minmax(sampleAllocations)
                                                        
}

