\name{scs-package}
\alias{scs-package}
\alias{scs}
\docType{package}
\title{
\packageTitle{scs}
}
\description{
\packageDescription{scs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scs}
\packageIndices{scs}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{scs}

Maintainer: \packageMaintainer{scs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
